+++ otterdog/credentials/__init__.py
@@ -5,7 +5,7 @@
 # which is available at https://spdx.org/licenses/MIT.html
 # SPDX-License-Identifier: MIT
 # *******************************************************************************
-
+import time
 from abc import abstractmethod
 from typing import Protocol, Optional
 
@@ -24,11 +24,19 @@ class Credentials:
         self.password = password
         self.github_token = github_token
         self._totp_secret = totp_secret
+        self._last_totp = None
 
     def get_totp(self) -> str:
-        totp = mintotp.totp(self._totp_secret)
-        utils.print_trace(f"generated totp '{totp}'")
-        return totp
+        while True:
+            totp = mintotp.totp(self._totp_secret)
+            utils.print_trace(f"generated totp '{totp}'")
+
+            if self._last_totp is None or totp != self._last_totp:
+                self._last_totp = totp
+                return totp
+            else:
+                utils.print_info("waiting 3s till generating new totp ...")
+                time.sleep(3)
 
     def __str__(self) -> str:
         return "Credentials(username={})".format(self.username)